fix kib c dan ubah id_ menjadi decimal

+++ tabel.sql
@@ -66,7 +66,7 @@ CREATE TABLE `data_mapping_rek_f` (
 
 CREATE TABLE `data_laporan_kib_a` (
   `id` int(11) NOT NULL auto_increment,
-  `id_tanah` double(20, 0) DEFAULT NULL,
+  `id_tanah` decimal(30, 0) DEFAULT NULL,
   `kode_skpd` varchar(255) DEFAULT NULL,
   `nama_skpd` varchar(255) DEFAULT NULL,
   `kode_lokasi` varchar(255) DEFAULT NULL,
@@ -102,7 +102,7 @@ CREATE TABLE `data_laporan_kib_a` (
 
 CREATE TABLE `data_laporan_kib_b` (
   `id` int(11) NOT NULL auto_increment,
-  `id_mesin` double(20, 0) DEFAULT NULL,
+  `id_mesin` decimal(30, 0) DEFAULT NULL,
   `kode_skpd` varchar(255) DEFAULT NULL,
   `nama_skpd` varchar(255) DEFAULT NULL,
   `kode_unit` varchar(255) DEFAULT NULL,
@@ -154,7 +154,7 @@ CREATE TABLE `data_laporan_kib_b` (
 CREATE TABLE `data_laporan_kib_c` (
   `id` int(11) NOT NULL auto_increment,
   `kode_skpd` varchar(255) DEFAULT NULL,
-  `id_gedung` double(20, 0) DEFAULT NULL,
+  `id_gedung` decimal(30, 0) DEFAULT NULL,
   `nama_skpd` varchar(255) DEFAULT NULL,
   `nama_unit` text DEFAULT NULL,
   `kode_unit` varchar(255) DEFAULT NULL,
@@ -200,7 +200,7 @@ CREATE TABLE `data_laporan_kib_c` (
 
 CREATE TABLE `data_laporan_kib_d` (
   `id` int(11) NOT NULL auto_increment,
-  `id_jalan_irigasi` double(20, 0) DEFAULT NULL,
+  `id_jalan_irigasi` decimal(30, 0) DEFAULT NULL,
   `kode_skpd` varchar(255) DEFAULT NULL,
   `nama_skpd` varchar(255) DEFAULT NULL,
   `nama_unit` text DEFAULT NULL,
@@ -244,7 +244,7 @@ CREATE TABLE `data_laporan_kib_d` (
 
 CREATE TABLE `data_laporan_kib_e` (
   `id` int(11) NOT NULL auto_increment,
-  `id_aset_tetap` double(20, 0) DEFAULT NULL,
+  `id_aset_tetap` decimal(30, 0) DEFAULT NULL,
   `kode_skpd` varchar(255) DEFAULT NULL,
   `nama_skpd` varchar(255) DEFAULT NULL,
   `nama_unit` text DEFAULT NULL,
@@ -496,7 +496,7 @@ CREATE TABLE `data_laporan_aset_lain` (
   `akumulasi_penyusutan` double(20, 0) DEFAULT NULL,
   `nilai_buku` double(20, 0) DEFAULT NULL,
   `jumlah_barang` int(11) DEFAULT NULL,
-  `id_mesin` double(20, 0) DEFAULT NULL,
+  `id_mesin` decimal(30, 0) DEFAULT NULL,
   `created_at` datetime DEFAULT current_timestamp,
   `active` tinyint(4) DEFAULT 1,
  PRIMARY KEY(id)
